# docs/swagger.yaml
definitions:
  domain.Item:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
  domain.ItemRequest:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
info:
  contact: {}
paths:
  /items:
    get:
      description: 모든 아이템을 조회합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Item'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 아이템 전체 조회
      tags:
      - items
    post:
      consumes:
      - application/json
      description: 새로운 아이템을 등록합니다.
      parameters:
      - description: 생성할 아이템 정보
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/domain.ItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Item'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 아이템 생성
      tags:
      - items
  /items/{id}:
    delete:
      description: ID로 특정 아이템을 삭제합니다.
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Item'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: 아이템 삭제
      tags:
      - items
    get:
      description: ID로 특정 아이템을 조회합니다.
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Item'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: 아이템 단건 조회
      tags:
      - items
swagger: "2.0"